型情報修正

+++ database/migrations/2019_05_14_141233_create_tasks_table.php
@@ -15,7 +15,7 @@ class CreateTasksTable extends Migration
     {
         Schema::create('tasks', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('folder_id')->unsigned();
+            $table->BigInteger('folder_id')->unsigned();
             $table->string('title', 100);
             $table->date('due_date');
             $table->integer('status')->default(1);
